+++ repository/src/main/java/com/walmartlabs/concord/repository/LastModifiedSnapshot.java
@@ -24,6 +24,7 @@ import com.walmartlabs.concord.common.FileVisitor;
 
 import java.io.IOException;
 import java.nio.file.Files;
+import java.nio.file.LinkOption;
 import java.nio.file.Path;
 import java.nio.file.attribute.BasicFileAttributes;
 import java.nio.file.attribute.FileTime;
@@ -50,6 +51,11 @@ public class LastModifiedSnapshot implements Snapshot, FileVisitor {
         if (prev == null) {
             return true;
         }
+
+        if (!Files.isRegularFile(path, LinkOption.NOFOLLOW_LINKS)) {
+            return false;
+        }
+
         return !prev.equals(attrs.lastModifiedTime());
     }
 }